fix: widen the hub's public stream parameter to intptr_t + bump cpu/cuda pins (#50)
A `cudaStream_t` is a pointer and therefore 64-bit, but every public
`stream` parameter in the hub's device-dispatch functions was declared
`int`. `ff::dt_euclidean` and friends inspect the DLTensor's device and
forward the same argument to either `FF_CPU::` or `FF_CUDA::`:

    if (is_cuda) return FF_CUDA::dt_euclidean(inp_out, voxel_spacing, stream);
    else         return FF_CPU::dt_euclidean(inp_out, voxel_spacing, stream);

so both backends must agree on the parameter's type. Widens 114 parameter
declarations across the 8 modules to `intptr_t`, matching the now-widened
cpu/cuda dispatch layers. This is a signature-only change (verified with the
same non-signature-line grep used in the cpu-lib/cuda-lib PRs: empty).

Bumps `cpu` -> d969e2d (fastfields-cpu-lib#61) and `cuda` -> 33aa9b3
(fastfields-cuda-lib#35).

Verified with `make all CXX=clang++`: the full hub (cpu library + hub object
files + link) builds successfully end to end against the bumped pins,
producing libfastfields.so linked against libfastfields-cpu.so. This is a
real, complete build of the CPU path; the CUDA path is not linked by this
target (CUDA needs `USE_CUDA=1` and nvcc, exercised at the cuda-lib layer
instead).

Refs fastfields/fastfields-lib#4

+++ distance.h
@@ -1,6 +1,7 @@
 #ifndef FF_LIB_DISTANCE
 #define FF_LIB_DISTANCE
 #include "dlpack.h"
+#include <cstdint>
 #include "defines.h"
 
 FF_NAMESPACE_BEGIN(FF)
@@ -48,7 +49,7 @@ FF_NAMESPACE_END(spline_t)
 void dt_euclidean(
           DLTensor & inp_out,
           double     voxel_spacing,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 /**
@@ -65,7 +66,7 @@ void dt_euclidean(
 void dt_l1(
           DLTensor & inp_out,
           double     voxel_spacing,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 /**
@@ -89,7 +90,7 @@ void dt_spline_table(
     const DLTensor & times,
           int8_t     spline = spline_t::Cubic,
           int8_t     bound  = bound_t::DCT2,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 /**
@@ -117,7 +118,7 @@ void dt_spline_brent(
           double     step,
           int8_t     spline = spline_t::Cubic,
           int8_t     bound  = bound_t::DCT2,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 /**
@@ -143,7 +144,7 @@ void dt_spline_gaussnewton(
           double     tol,
           int8_t     spline = spline_t::Cubic,
           int8_t     bound  = bound_t::DCT2,
-          int        stream = 0
+          intptr_t   stream = 0
 );
 
 /**
@@ -166,7 +167,7 @@ void dt_mesh(
     const DLTensor & faces,
           bool       _signed = true,
           bool       naive   = false,
-          int        stream  = 0
+          intptr_t   stream  = 0
 );
 
 FF_NAMESPACE_END(FF)
